+++ baybe/surrogates/gaussian_process/presets/core.py
@@ -11,6 +11,9 @@ class GaussianProcessPreset(Enum):
     BAYBE = "BAYBE"
     """The default BayBE settings of the Gaussian process surrogate class."""
 
+    BOTORCH = "BOTORCH"
+    """The BoTorch settings."""
+
     CHEN = "CHEN"
     """The adaptive kernel hyperprior settings proposed by :cite:p:`Chen2026`."""